requirements.txt
pandas==2.0.3
scikit-learn==1.2.2
pickle==4.0
apache-beam
pandas
pytest
joblib
fastapi[all]
pydantic
pyyaml
